Release 1.4.5

+++ learn-puppet/src/main/resources/scripts/applyPuppetRemote.sh
@@ -4,8 +4,12 @@
 # PACKAGE_NAME
 # DEPLOYMENT_HOST
 # SSH_USER
-# LIFERAY_HOME
 # SRC_ARC
+#
+# an optional variable:
+# LIFERAY_HOME
+
+PROJECT="nkeservice"
 
 if [ $SRC_ARC ] ; then
     echo "SRC_ARC is $SRC_ARC."
@@ -18,15 +22,16 @@ PRGDIR=`dirname "$0"`
 
 if [ $LIFERAY_HOME ] ; then
     echo "LIFERAY_HOME is $LIFERAY_HOME."
-else
-    echo "LIFERAY_HOME is not defined."
-    exit 1
-fi
 
-LIFERAY_DEPLOY_DIR="$LIFERAY_HOME/deploy/"
-export LIFERAY_DEPLOY_DIR
-LIFERAY_TOMCAT_ROOT="$LIFERAY_HOME/tomcat-7.0.27"
-export LIFERAY_TOMCAT_ROOT
+    LIFERAY_DEPLOY_DIR="$LIFERAY_HOME/deploy/"
+    export LIFERAY_DEPLOY_DIR
+    LIFERAY_TOMCAT_ROOT="$LIFERAY_HOME/tomcat-7.0.27"
+    export LIFERAY_TOMCAT_ROOT
+
+    echo "LIFERAY_DEPLOY_DIR is $LIFERAY_DEPLOY_DIR."
+else
+    echo "WARNING: LIFERAY_HOME is not defined."
+fi
 
 if [ $PACKAGE_NAME ] ; then
     echo "PACKAGE_NAME is $PACKAGE_NAME."
@@ -49,35 +54,38 @@ else
     exit 1
 fi
 
-if [ $LIFERAY_DEPLOY_DIR ] ; then
-    echo "LIFERAY_DEPLOY_DIR is $LIFERAY_DEPLOY_DIR."
+if [ $# -eq 1 ]; then
+    MANIFEST="$1"
 else
-    echo "LIFERAY_DEPLOY_DIR is not defined."
-    exit 1
+    MANIFEST="manifests/site.pp"
 fi
 
 PATH=$PATH:/opt/csw/bin/
 export PATH
 
-scp $SRC_ARC $SSH_USER@$DEPLOYMENT_HOST:/tmp/$PACKAGE_NAME.zip &&
+# if we deploy Liferay, archive with EE licenses should be copied to the target machine first.
+if [ $LIFERAY_HOME ]; then
+    scp $SRC_ARC $SSH_USER@$DEPLOYMENT_HOST:/tmp/$PACKAGE_NAME.zip &&
 
-ssh $SSH_USER@$DEPLOYMENT_HOST "PATH=\$PATH:/opt/csw/bin/;
+    ssh $SSH_USER@$DEPLOYMENT_HOST "
 
-sudo rm -rf /tmp/learn-deploy/; sudo mkdir /tmp/learn-deploy/ &&
+        sudo rm -rf /tmp/$PROJECT-deploy/; sudo mkdir /tmp/$PROJECT-deploy/ &&
 
-sudo unzip -o /tmp/$PACKAGE_NAME.zip -d /tmp/learn-deploy/ &&
+        sudo unzip -o /tmp/$PACKAGE_NAME.zip -d /tmp/$PROJECT-deploy/ license*.xml &&
 
-sudo rm -f /tmp/$PACKAGE_NAME.zip &&
+        sudo rm -f /tmp/$PACKAGE_NAME.zip &&
 
-sudo puppet apply --trace -v --modulepath=/tmp/learn-deploy/modules/:/etc/puppet/modules/:/usr/share/puppet/modules/ /tmp/learn-deploy/learn-liferay.pp" #&&
+        sudo puppet agent -v --onetime --no-daemonize 
+    "
+    if [ $? -ne 0 ]; then
+        echo "Puppet run failure on $DEPLOYMENT_HOST"
+        exit 1
+    fi
+else
+    ssh $SSH_USER@$DEPLOYMENT_HOST  "sudo puppet agent -v --onetime --no-daemonize"
+    if [ $? -ne 0 ]; then
+        echo "Puppet run failure on $DEPLOYMENT_HOST"
+        exit 1
+    fi
+fi
 
-#/bin/bash $PRGDIR/testLiferayAvailable.sh # &&
-
-#ssh $SSH_USER@$DEPLOYMENT_HOST "PATH=\$PATH:/opt/csw/bin/; ENV=\`sudo puppet config print environment\` && sudo find /tmp/learn-deploy/licenses/\$ENV -type f -name 'license*.xml' | sudo xargs -I files cp files $LIFERAY_DEPLOY_DIR &&
-#    sudo find /tmp/learn-deploy/files -type f -name '*.war' | sudo xargs -I files cp files $LIFERAY_DEPLOY_DIR && sudo ls -l $LIFERAY_DEPLOY_DIR" # &&
-
-#/bin/bash $PRGDIR/testDeploymentFinished.sh &&
-
-#scp $PRGDIR/applyPostDeployConfiguration.sh $SSH_USER@$DEPLOYMENT_HOST:/tmp/ &&
-
-#ssh $SSH_USER@$DEPLOYMENT_HOST "PATH=\$PATH:/opt/csw/bin/; sudo LIFERAY_TOMCAT_ROOT=$LIFERAY_TOMCAT_ROOT /bin/bash /tmp/applyPostDeployConfiguration.sh"
